[IMP] Separating common features in test into a different file
