+++ product_unique_serial/__openerp__.py
@@ -4,13 +4,13 @@
     'author': "Vauxoo",
     'website': "http://www.vauxoo.com",
     'category': 'stock',
-    'version': '1.0',
+    'version': '8.0.0.1.0',
     'license': "AGPL-3",
     'depends': ['stock_no_negative'],
     'data': [
         "views/product_view.xml",
         "views/stock_view.xml",
-        "views/stock_serial_view.xml"
+        "views/stock_serial_view.xml",
     ],
     'demo': [
         "demo/test_demo.xml",
