+++ product_unique_serial/i18n/es_PA.po
@@ -77,9 +77,7 @@ msgstr "ID"
 #. module: product_unique_serial
 #: help:stock.picking.type,use_create_lots:0
 msgid "If this is checked only, it will suppose you want to create new Serial Numbers / Lots, so you can provide them in a text field. "
-msgstr ""
-"SI este campo es seleccionado, supondrá que quieres crear nuevos Números  de "
-"Serie, entonces tu puedes proporcionarlos en un campo de texto. "
+msgstr "Si este campo es seleccionado, supondrá que quieres crear nuevos Números Seriales / Lotes, entonces tu puedes proporcionarlos en un campo de texto. "
 
 #. module: product_unique_serial
 #: help:stock.picking.type,use_existing_lots:0
@@ -96,7 +94,7 @@ msgstr "Línea inventario"
 #. module: product_unique_serial
 #: field:stock.move,lot_unique:0
 msgid "Is lot unique?"
-msgstr "¿Lote Único?"
+msgstr "¿Es lote único?"
 
 #. module: product_unique_serial
 #: field:stock.production.lot,last_location_id:0
@@ -221,7 +219,7 @@ msgstr "Serial de Productos"
 #: code:addons/product_unique_serial/wizard/stock_serial.py:196
 #, python-format
 msgid "Serial {} not found"
-msgstr "Numero de Serial {} No Encontrado"
+msgstr "Número de serial {} no encontrado"
 
 #. module: product_unique_serial
 #: field:stock.serial,serial_ids:0
@@ -247,7 +245,7 @@ msgstr "El numero serial {} ya fue capturado"
 #. module: product_unique_serial
 #: model:ir.model,name:product_unique_serial.model_stock_picking_type
 msgid "The picking type determines the picking view"
-msgstr "El tipo de Albarán determina la vista del albarán"
+msgstr "El tipo de albarán determina la vista del albarán"
 
 #. module: product_unique_serial
 #: code:addons/product_unique_serial/model/stock.py:96
@@ -286,14 +284,12 @@ msgstr "Advertencia"
 #: code:addons/product_unique_serial/model/product.py:43
 #, python-format
 msgid "You can't activate 'full lots traceability' or 'unique lot' in the product %s because this product already have stock movements."
-msgstr ""
-"No se puede activar 'Trazabilidad completa de lotes' o 'Lote unico' en el "
-"producto %s porque este producto ya tiene movimientos de existencias."
+msgstr "No se puede activar 'trazabilidad completa de lotes' o 'lote unico' en el producto %s porque este producto ya tiene movimientos de existencias."
 
 #. module: product_unique_serial
 #: view:stock.serial:product_unique_serial.view_stock_serial
 msgid "of"
-msgstr "De"
+msgstr "de"
 
 #. module: product_unique_serial
 #: view:stock.serial:product_unique_serial.view_stock_serial
