+++ product_unique_serial/tests/__init__.py
@@ -1,2 +1,3 @@
+from . import test_common
 from . import test_for_unicity
 from . import test_serial_wizard
