+++ product_unique_serial/tests/test_for_unicity.py
@@ -24,136 +24,20 @@
 #
 ###############################################################################
 
-from copy import deepcopy
-
 from openerp import exceptions
-from openerp.tests.common import TransactionCase
 from openerp.tools import mute_logger
 from psycopg2 import IntegrityError
-from openerp import _
+from .test_common import TestCommon
 
 
-class TestUnicity(TransactionCase):
+class TestUnicity(TestCommon):
 
-    """
-    This test will prove the next cases to procure the
-    module unicity:
+    """ This test will prove the next cases to procure the module uniqueness:
     - Test 1: Can't be created two Serial Numbers with the same name
     """
 
-    note = _(
-        u'Remember: When a serial number (lot) is selected, its quantity '
-        u'is fixed against the quantity in the serial number (lot) and not '
-        u'against the quantity in full of the product.')
-
-    msg_greater = _(
-        u'Product %s has been configured to use unique lots. '
-        u'You are trying to set %s items in lot %s. %s')
-
-    msg_increase = _(
-        u'Product %s has been configured to use unique lots. '
-        u'You are trying to increase %s items in the lot %s. %s')
-
-    def setUp(self):
-        super(TestUnicity, self).setUp()
-        self.stock_production_lot_obj = self.env['stock.production.lot']
-        self.stock_picking_type_obj = self.env['stock.picking.type']
-        self.stock_picking_obj = self.env['stock.picking']
-        self.product_obj = self.env['product.product']
-        self.stock_move_obj = self.env['stock.move']
-        self.product_uom_obj = self.env['product.uom']
-        self.stock_location_obj = self.env['stock.location']
-        self.stock_inventory_obj = self.env['stock.inventory']
-        self.stock_inventory_line_obj = self.env['stock.inventory.line']
-        self.stock_loc = self.env.ref('stock.stock_location_stock')
-        self.prod_d1 = self.env.ref('product_unique_serial.product_demo_1')
-        self.stock_production_lot_obj = self.env['stock.production.lot']
-
-    def create_stock_picking(self, moves_data, picking_data, picking_type):
-        """ Returns the stock.picking object, with his respective created
-            created stock.move lines """
-        # Require deepcopy for clone dict into list items
-        moves_data_copy = deepcopy(moves_data)
-        picking_data_copy = picking_data.copy()
-        stock_move_ids = []
-        default_product_data = None
-        for move_n in moves_data_copy:
-            # Getting the qty for the stock.move
-            qty = move_n.get('qty')
-            del move_n['qty']
-            # Getting default data through product_id onchange
-            if 'source_loc' in move_n.keys()\
-                    and 'destination_loc' in move_n.keys():
-                default_product_data = self.stock_move_obj.onchange_product_id(
-                    prod_id=move_n.get('product_id'),
-                    loc_id=move_n.get('source_loc'),
-                    loc_dest_id=move_n.get('destination_loc'))
-                del move_n['source_loc']
-                del move_n['destination_loc']
-            else:
-                default_product_data = self.stock_move_obj.onchange_product_id(
-                    move_n.get('product_id'))
-            move_n.update(default_product_data.get('value'))
-            # Getting default data through product_uom_qty onchange
-            default_qty_data = self.stock_move_obj.onchange_quantity(
-                move_n.get('product_id'),
-                qty,
-                default_product_data.get('product_uom'),
-                default_product_data.get('product_uos'))
-            default_qty_data['value'].update({'product_uom_qty': qty})
-            move_n.update(default_qty_data.get('value'))
-            # Getting the new stock.move id
-            move_created = self.stock_move_obj.with_context(
-                default_picking_type_id=picking_type.id).create(move_n)
-            stock_move_ids.append(move_created.id)
-        # Creating picking
-        picking_data_copy.update({
-            'move_lines': [(6, 0, stock_move_ids)],
-            # TODO: Uncomment when fix current context bug
-            # 'move_lines': [(0, 0, moves_data_copy)],
-            # and remove move_created line
-            'picking_type_id': picking_type.id})
-        return self.stock_picking_obj.create(picking_data_copy)
-
-    def transfer_picking(self, picking_instance, serial_number=None):
-        """ Creates a wizard to transfer the picking given like parameter """
-        # Marking the picking as Todo
-        picking_instance.action_confirm()
-        if picking_instance.state == 'confirmed':
-            # Checking availability
-            picking_instance.action_assign()
-        if picking_instance.state == 'assigned':
-            # Transfering picking
-            transfer_details = picking_instance.do_enter_transfer_details()
-            wizard_for_transfer = self.env[transfer_details.get('res_model')].\
-                browse(transfer_details.get('res_id'))
-            if serial_number:
-                if len(serial_number) == 1:
-                    for transfer_item in wizard_for_transfer.item_ids:
-                        transfer_item.lot_id = serial_number[0].id
-                else:
-                    # Case for the A, B & C serial numbers: That are 3 like
-                    # value for the quantity field in the wizard, so we should
-                    # split this record 2 times.
-                    # Splitting the record and executing the transfer
-                    wizard_split = wizard_for_transfer.item_ids[0].\
-                        split_quantities()
-                    wizard_for_transfer = self.env['stock.transfer_details'].\
-                        browse(wizard_split.get('res_id'))
-                    wizard_split = wizard_for_transfer.item_ids[0].\
-                        split_quantities()
-                    wizard_for_transfer = self.env['stock.transfer_details'].\
-                        browse(wizard_split.get('res_id'))
-                    index = 0
-                    for transfer_item in wizard_for_transfer.item_ids:
-                        transfer_item.lot_id = serial_number[index].id
-                        index += 1
-            # Executing the picking transfering
-            wizard_for_transfer.do_detailed_transfer()
-
     def test_1_1_1product_1serialnumber_2p_in(self):
-        """
-        Test 1.1. Creating 2 pickings with 1 product for the same serial
+        """ Test 1.1. Creating 2 pickings with 1 product for the same serial
         number, in the receipts scope, with the next form:
         - Picking 1 IN
         =============================================
@@ -198,9 +82,8 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_1_2_1product_1serialnumber_2p_track_incoming(self):
-        """
-        Test 1.2. (track incoming) Creating 2 pickings with 1 product for the
-        same serial number, in the receipts scope, with the next form:
+        """ Test 1.2. (track incoming) Creating 2 pickings with 1 product for
+        the same serial number, in the receipts scope, with the next form:
         - Picking 1 IN
         =============================================
         || Product ||  Quantity  ||  Serial Number ||
@@ -252,8 +135,7 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_2_1_1product_1serialnumber_2p_out(self):
-        """
-        Test 2.1. Creating 2 pickings with 1 product for the same serial
+        """ Test 2.1. Creating 2 pickings with 1 product for the same serial
         number, in the delivery orders scope, with the next form:
         - Picking 1 OUT
         =============================================
@@ -307,9 +189,9 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_2_2_1product_1serialnumber_2p_track_outgoing(self):
-        """
-        Test 2.2. (track outgoing) Creating 2 pickings with 1 product for the
-        same serial number, in the delivery orders scope, with the next form:
+        """ Test 2.2. (track outgoing) Creating 2 pickings with 1 product for
+        the same serial number, in the delivery orders scope, with the next
+        form:
         - Picking 1 OUT
         =============================================
         || Product ||  Quantity  ||  Serial Number ||
@@ -370,9 +252,8 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_3_1product_qtyno1_1serialnumber_1p_in(self):
-        """
-        Test 3. Creating a picking with 1 product for the same serial number,
-        in the delivery orders scope, with the next form:
+        """ Test 3. Creating a picking with 1 product for the same serial
+        number, in the delivery orders scope, with the next form:
         - Picking 1 IN
         =============================================
         || Product ||  Quantity  ||  Serial Number ||
@@ -402,8 +283,7 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_4_1product_qty3_3serialnumber_1p_in(self):
-        """
-        Test 4. Creating a picking with 1 product for three serial numbers,
+        """ Test 4. Creating a picking with 1 product for three serial numbers,
         in the receipts scope, with the next form:
         - Picking 1
         =============================================
@@ -441,9 +321,8 @@ class TestUnicity(TransactionCase):
         )
 
     def test_5_1product_1serialnumber_2p_internal(self):
-        """
-        Test 5. Creating 2 pickings with 1 product for the same serial number,
-        in the internal scope, with the next form:
+        """ Test 5. Creating 2 pickings with 1 product for the same serial
+        number, in the internal scope, with the next form:
         - Picking 1 INTERNAL
         =============================================
         || Product ||  Quantity  ||  Serial Number ||
@@ -506,9 +385,7 @@ class TestUnicity(TransactionCase):
 
     @mute_logger('openerp.sql_db')
     def test_6_1serialnumber_1product_2records(self):
-        """
-        Test 6. Creating 2 identical serial numbers
-        """
+        """ Test 6. Creating 2 identical serial numbers """
         product_id = self.env.ref('product_unique_serial.product_demo_1')
         lot_data = {
             'name': '86137801852520',
@@ -521,11 +398,8 @@ class TestUnicity(TransactionCase):
             self.stock_production_lot_obj.create(lot_data)
 
     def test_7_1_1product_1serialnumber_track_production_in(self):
-        """
-        Test 7. Creating moves as production order with 1 product
-        as material, 2 moves with 1 qty  and 1 same serial number
-        for both
-        """
+        """ Test 7. Creating moves as production order with 1 product as
+        material, 2 moves with 1 qty  and 1 same serial number for both """
         product = self.env.ref('product_unique_serial.product_demo_1')
         # track_production and lot_unique_ok to test unicity
         self.assertTrue(product.write({'track_all': False,
@@ -565,10 +439,8 @@ class TestUnicity(TransactionCase):
         self.assertEquals(err.exception.value, msg)
 
     def test_7_2_1product_1serialnumber_track_production_out(self):
-        """
-        Test 7.2. Creating moves as finished product, 1 product,
-        2 moves with 1 qty and 1 same serial number for both
-        """
+        """ Test 7.2. Creating moves as finished product, 1 product, 2 moves
+        with 1 qty and 1 same serial number for both """
         product = self.env.ref('product_unique_serial.product_demo_2')
         # track_incoming, track_prodcution and lot_unique_ok to test unicity
         self.assertTrue(product.write({'track_all': False,
